refactor(runtime): decompose template/lifecycle.rs into a lifecycle/ module tree
